Move print table to central location for reuse
